Add tests and fixes for GSKColor

+++ Classes/GSKColor.h
@@ -11,14 +11,14 @@
 @property (nonatomic) NSInteger red;
 @property (nonatomic) NSInteger blue;
 @property (nonatomic) NSInteger green;
-@property (nonatomic) NSInteger hue;
-@property (nonatomic) NSInteger saturation;
-@property (nonatomic) NSInteger value;
+@property (nonatomic) CGFloat hue;
+@property (nonatomic) CGFloat saturation;
+@property (nonatomic) CGFloat value;
 
 // Creating colors
 - (id)initWithARGB:(NSUInteger)argb;
 - (id)initWithAlpha:(NSInteger)alpha red:(NSInteger)red green:(NSInteger)green blue:(NSInteger)blue;
-- (id)initWithAlpha:(NSInteger)alpha hue:(NSInteger)hue saturation:(NSInteger)saturation value:(NSInteger)value;
+- (id)initWithAlpha:(NSInteger)alpha hue:(CGFloat)hue saturation:(CGFloat)saturation value:(CGFloat)value;
 
 // Instance methods
 - (NSUInteger)gl;
